dashboard/requirements_dashboard.txt
streamlit
pandas
psycopg2-binary
python-dotenv
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
plotly
altair
bcrypt
